Added a basic panel for the scrolling credits after escaping
Once the player reaches the outside, a credits panel will start to enter their view on the menu, even when the menu is in the empty state. The menu slides up into view as the player spends time in the outside and while in non-clickable menus, like the video panel or transitionning in or out of other panel menus.