Added more animations to the credits panel
The credits panel is now fully animated when going from different states in the menu. The panel will also disable itself when it is not in view, such as in the video or sensitivity states.

Also, previous uses of anchoredPosition have been updated to use vector2s instead of vector3s.