Added a slow-mo effect when the player breaks the glass
When the player breaks the glass of the startingRoom's window, the game slows time by reducing the speed at which the player moves, the velocity and rotation of the window's broken glass particles and the pitch of any audio that would be used at the time.

A few issues need to be fixed, such as the player clearly teleports as the camera jumps some distance upon teleporting. Also, we can see rendering issues with the outside portal when looking at it in the edges of the screen from affar.