Simplified how the player handles movement
The HandlePlayerMovement function now handles each movement vector one at a time. The previousPosition tracking has been removed.

The bug we are trying to fix is as follows: Whenever the player passes through the startingRoom's window, and whenever the player passes through their first portal since startup, they will get pushed forward a noticable distance. After rigorous testing, it has been discovered that is is due to the player's collider, as the extra movement is done between the end of fixedupdate and the start of Update. Disabling the player's collider will prevent this issue from occuring for both scenarios.

Possible Idea: It might be the collision from the portal's meshCollider when it is enabled. When the player collides with an object, look to see what it is.