Fixed the bug where the player was colliding with the portals
The player was previously colliding with portals, causing them to be noticably moved after teleporting from either their first portal or the startingRoom's window. Now, all portals use a trigger boxCollider for testing it's edges instead of a meshCollider, which was causing the issue.

However, there is still the issue of the portal outside rendering wierdly when looking back. IS it due to the windows using the exterior side of the portals?