Fixed the rendering issue of the outside window's portal when on the screen edges
The issue that was causing the portal to sometimes dissapear had to do with how it was detecting what edge was connecting to the planes representing the camera's fustrum.

This bug was fixed by including each edge of the portal to the cameraBoundsVerts array once they have been converted to a position relative to the camera's viewport space, even if they are out of bounds. Originally, a vertex that was out of bounds was put into a list which would ray trace and collide with the camera's fustrum planes. Currently, this is still implemented, but even when the vertex is found to be out of the camera's view, it will also be added to the cameraBoundsVerts list.