Improved many small elements of the game
Many  small adjustments have been made to the game, such as: Altered the glass shard texture and the amount emitted, the music fade timings for the intro and made the starting music use the muted source. Removed many debug calls.

Two remaining bugs that need to be fixed are when the player hits a hard landing, they pass through the floor and at times when looking at portals an error will occur involving something out of view, I cannot recall.