Refactored how the player handles steps and movement
Whenever the player recieves a request to move, it will handle it the moment it recieves it and move the player accordingly. This has fixed the bug involving the player not handling stepping properly, as it was sending a move request that ignored the previous move requests that were to re-position the player.