Created a new solution for the puzzleRoom 6
PuzzleRoom 6 has a new solution. It still wants you to fall into the first pit and move to a specific position while falling, but this time it leads the player back to the same side they started on. They can either skip part of it by falling back into the hole, getting to the platforming section, of they can land on the tunnel and enter the other side of the room. There, the wall portal will let them jump into the top-side's tunnel hole, bringing them to the top side again, bringing them to the platforming section.

However, when falling back into this room with the escape rotation, it is possible for the player to get flipped around if they fall into the tunnel entrance that it will flip their escape rotation and make them fall forward into puzzleRoom 7.