Added another portal set to the puzzleRoom 6
The new portal added flips the player when falling through the tunnel on the top side of the room. This will preserve the solution to the puzzle, only requiring the player to turn around (which gives a more non-eucledian feel to the room). This new portal however will allow the player to continue falling through the rooms, as now it will prevent the player from losing their escape angle when they fall into the tunnel on their way back.