Updated how the playerSounds handles looping
The plaerSounds will now start a new song when using muted songs in the puzzleRooms. The functionnality behind looping songs has been changed a bit, as now the outside sounds looping on their own is independent of the current state of the player, as long as the source is using the outsideSounds clip.