Each puzzleRoom now has it's own adjustable color
Each puzzleRoom can now adjust the color tint applied to it's wall shader. The wallShader now accepts a vector3 to use as the color, which controls how the walls of the puzzleRoom look up until the above and bellow extremes of the room. This has allowed the walls to have a gradient effect when falling, and it keeps the teleport of the player seamless.

Also, the tint can be set to (0, 0, 0) so that no extra color is applied. However, the texture will still be forced into black and white. This may be changed later.