Added a color for each room and adjusted the wall shader some more
Each room now has a unique color used for it's walls. The next update will feature new colored textures that reflect this color theme for the room.

Removed the conversion to grayscale done when applying the tint to the wall texture, along with changing the canvas texture applied to all the walls to a grayscaled version.

Also, fixed the glitch where the pixel height was not properly being set. Now we convert the vertex to their world position before getting the height.