Added rotation, starting and ending to the loading scene
The loadingScene's boxes now have a rotation element added to the boxes, which also uses a sepperate step count than the position animation.

The animation also stars with the boxes off-screen, makign them slide into view. Once the game loads (in this case, the right keypress), the boxes will slide out of view to leave the screen white again.