Adde another box and component to the animation
Another box is added to the animation so that the center box did not share the same rotation and position step amounts, making the scene more distracting as the game loads.

Added a height component to the positional animation: The further into loop an animation is, the larger a bonus height is applied. This makes it so when the cubes go back down for a quick dip, they go lower than their normal height. It has a very cool effect that I am proud of making because of how simple it is.

Also, each box now has another child box that is placed between the parent box and the camera. This box is colored white and is slightly smaller, making it seem like the parent box is hollow.