Improved the scene transitions and loading process
The loading scene now properly loads into the game scene by setitng allowSceneActivation to false. This way the scene will detect when the loading is nearly finished, then go thorugh with the scene load.

Also, since we are now tracking the coroutine and it's progress, the cubes will now be used to indicate the loading progress of the scene. The child white cube of each loading cube will have it's size be adjusted relative to the loading progress, which each white cube being invisibly small once the loading is complete.

When debugging the loading process, it is revealed that the game spends most of it's time loading in the progress range of [0.01, 0.02], so when meassuring the progress it is adjusted to visualize the progress of the loading around that range.