Improved a couple puzzleRoom's portals and added a startup animation for the loading bar
The loading bar that is used to meassure the loading progress of the terrain generation has an animation that plays relative to the timing of since this scene loaded. It animates the loading bar up into position and from fully opaque to fully visible.

PuzzleRoom 3 had one of it's pillars repositionned so that teleporting will not cause unalligned textures to be noticable on teleport.

PuzzleRoom 5 now has another layer of cubes around each portal to prevent the wierd rendering issue that let the player camera often view past the portal and the cubes around it's edges.