Converted large audio files from wav to mp3.
converted all the large .wav audio files to .mp3 to preserve the size of all used assets. Personally, I cannot hear the difference, but I would not call myself an audiophile.

Also, when the player camera collides with a collider, it will now go back a set amount instead of a portion of the camera's height. This makes it so the camera seems stationnairy when jumping into a ceiling.