Moved the terrian, startingRoom and player into the loading scene.
The terrain generation will be the first thing to load before trying to load the puzzleRooms (maybe even load them at the same time?). The idea is for the terrain to be visible as soon as possible so there is a focus on the outside as the game loads.

Also, the loading animation is being converted into the menu script by changing the boxes to be elements in the UI.