New scene will now load using the menu
The menu now loads the scene that contains all the puzzleRooms. It controls the state of the start button and will now let the player play the game using this system.

Many small fixes to adjusting the game as it uses two scenes, as now it uses two scenes new bugs have come up.