Added the choice of a second tint for the puzzleRoom walls. PuzzleRoom 6 updated.
Each puzzleRoom now has an alt tint and a center offset. The alternative tint is used on the bottom-half of the puzzleRoom, and mainly exists for puzzleRoom 6.

PuzzleRoom 6 has it's returningWall portal adjusted so that when jumping into the returning wall portal, the goal of jumping into the tunnel is much more obvious.