PuzzleRooms now set the textures of it's materials dynamically
Other than the walls and clouds materials, the puzzleRooms now set the textures and alternative textures dynamically using the textures set in the script.

Each puzzleRoomEditor now has an array of materials and textures. This serves the same purpose of setting the textures and alt textures to the floor and column materials, but it is done so using dynamically sizes arrays. This means we are not forced to always use a pillar material, and we are able to use multiple materials and textures for multiple blocks.