Implemented the second set of floor textures in puzzleRoom 6
The puzzleRoom6 now has another set of materials used for the bottom set of the room. They are linked to the bottom materials which are using temp texture values.

The puzzleRoom has been adjusted to make the path more clear: The starting path from the entrance to the tunnel is connected, along with the path from the exit and the bottom tunnel. Also, the top tunnel has been shortened.

The changes comming to this room will fixed the textures and make the textures seamlessly connect between the floors with the portals. To make the best of this effect, we may need to introduce another set of floor materials.