Added a new custom shader to the waitingRooms, adding a color tint blend
Each waitingRoom now requires two given color tints in the editor that control a color tint that fades from the waitingRoom's material to the given tint in the two attachedRooms. This color tint should be set to the same color tint as the puzzleRoom, but that doesn't seem to be fully accurate as the wall textures are muhc lighter than the waitingRoom's textures.