WaitingRoom's windowFrame material now uses the custom shader
The windowFrame material that is also used on  the doorframe boxes now uses the new custom shader that implements a color tint. This change does not make the doorframe more noticable as it now blends into the two rooms. I do not know if this is better than the stark contrast from using the default wood texture.

Also, the ceiling, floor and wall all use a different amount of the color tints, and each waitingRoom has had their color tint adjusted. Simply using the value from the puzzleRoom is not the perfect solution and requires a bit of hand-tunning