Added a DetectPlayerLegRay to windows
The interior waitingRooms now add a DetectPlayerLegRay script to their interior windows. The use of this is to change the texture used on the window from a normal glass texture to a cracked glass texture. This change will occur when the player's leg comes in contact with the window and they are falling past a certain limit. This is to let the player crack a window when landing on it, giving them the hint to go back to the startingRoom and break the pre-cracked glass.