Updated the window's particleEmitter
The window's particle emitter now takes into account the collider of the waitingRoom's floor. The properties of the particles are also updated, using the glass shard material and using random velocities.

I am having trouble trying to set a random starting value for the rotation and size for the particles however.