Adjusted how the particleSystem used on the windows is created. Changed player, terrain and menu settings to be build-ready
The particleSystem used for the windows is now preset at the start in a reference object  that each waitingRoom has a reference to. When a window's glass object is created, it uses said reference so that the particlesystem is easily duplicated. This also allowed us to change the starting values of the particlesystem's main, such as using curves for starting size.

The menu now properly meassures the progress of the scene loading and the boxes will now fully leave the screen.

The player now regains they timeRate slower and the timeRate will start to increase if they are either very far from the window or looking up into the window as they are falling.