Moved around waitingRooms, minor sounds and puzzleRoom 3 update.
Outside music now has it's loop property set to true so that the background effects clearly loop instead of replaying the source. This however requires us to set the upgraded source back to non-looping once a new song is played.

Adjusted the sizes of eahc waitingRoom and their attachedRooms so that once the player enters the proper escape angle, its inevitable for them to land onto a glass window.

Also changed the flooring of puzzleRoom 3 so that the player notices much easier that the room has flipped.