Loading boxes are now independent from eachother
Each loadingBox now has their own boolean and float to track their loading progress. Once a box is filled, it will lower off the screen. This also allows each box to rise into view on sepperate values.

The animation used on the buttons when entering the video panel has been changed to fix the issue where the buttons locked to the screen edge. This may change again with a better video panel animation in general.

Also, by holding down a key during startup from 1-0 or q-p from the keyboard, it will change the boxCount used when loading (to test which box count is the most desired)