Widened the top of the tunnel in puzzleRoom 6
The tunnel entrance on the top part of the puzzleRoom 6 has been widened. That way, when the player is falling into the top part of the room from the clouds, there is an obvious visual indication of where to land if the player wants to land on top of the tunnel where the required portal is placed.