Added a hint panel in the menu
A hint panel is added to the menu that will be displayed any time the menu is open. It is used to give the player hints to how the game controls. So far, it informs the player that they are able to reset. The text can be controlled via a script and uses an array of strings, taking in an index to determine what to change the text to.