Hint panel now informs the player about running and jump priming
The hint panel now update's it's text depending on the puzzleRoom the player is in. This will inform the player how to run and that they are able to perform jumps using the space bar, and can even prime their jumps.

The hint panel can now slide into and out of view on it's own, even when the menu is closed. This is how the game will inform the player of how to run if they have not yet figured it out.