Added a hack to fix the "screen position out of view fustrum" bug
An  "incompatible" boolean has been added to the portalSet script. A portal set that is  "incompatible" will not render another portalSet if it is also  "incompatible".  All waitingRoom portals are preset to be  "incompatible". By setting the portals which are prone to causing the out of fustrum error to  "incompatible", then we can avoid the bug without worrying about not rendering a portal.

This hack will however require more testing to find more problem portals.