Changed the "screen position" hack into a proper fix
The bug was discovered to be an issue with the camera's view size. It would try to get a position in the camera's screen when the camera's height or width pixel size was at 0.

By having the viewingCamera's pixel size checked before running CalculateViewingRect(), we can catch this error and return out of RenderCameraView safely.

Also Altered the waitingRoom's window sizes to prevent the player from clipping into the walls.