Updated Purple room (PuzzleRoom 7)
The purple room has been updated. One of the known paths that lead to the player skipping the red/final room has been changed to prevent the player from reaching the escape rotation. Also, the red room had it's floor updated, making it so the player cannot enter it on the default rotation.