Moved objects and adjusted blending of the blocks in puzzleRoom 7
Moved cubes, portals and stairs around in the room and changed their blending amounts to smooth out the blocks. The room is now in a finished state, but a few changes may be added if needed.

Also when the player falls back into the startingRoom from the outside, their camera will start re-rendering the puzzleRooms and not the outside terrain.