Adjusted red room to be capable of putting the player in the proper rotation even from the default rotation
The final red puzzleRoom has it's portals and blocks adjusted so that it is capable of properly setting the player to the escape rotation even when entering at the default rotation. This is because the previous puzzleRoom has been adjusted so that it's easy to leave the room on the default rotation.

Also connected parts of puzzleRoom 7 together and added depth to most objects to connect them to walls instead of having them slightly attached to other parts of the room.