Swapped the colors and textures of room 1, 4 and 7
The reason behind texture and color swapping is so that puzzleRoom 7 becomes the black/white textured room. This was accomplished by moving room 7's textures into 4, r's textures into 1 and 1's textures into 7.

Also, the blend effect on puzzleRoom 6 was changed to cover the entirety of the walls. Also, the second wall texture was set to be the canvas, just like the first texture. These changes look slightly better with the normal textures, but using the alternative textures makes the walls look much better as the texture changes throughout the wall and the player can see how large the room is. It's actually really cool.