Added a run toggle. It is conveyed to the player through the hint messages.
Added a "run toggle" mechanic. Pressing left-control will toggle the player to always run, making shift cause the player to walk instead.

Slightly changed the player's walking and running speed. They walk faster now, and running makes them run a small amount faster than the previous speed settings.

Changed prime jump to normal jump. The playercontroller now tracks when they have entered the puzzleRoom 3, and whether they have jumped once they have entered the room. This is so that the player has an incentive to jump off the platform by making the hint appear when entering the room even if they have already jumped in previous rooms. And, the room itself has been adjusted to plant the idea in the player's head to jump.

Also, the puzzleRoom hint setting has been pushed into the playerController for simplicity as having it all done in the puzzleRoomEditor script was not it's ideal location. However, the call will still be done from the puzzleRoomEditor script as it is the easiest way to determine what room the player is in. And, by tracking the player's room number, we can skip showing hints for rooms that the player has back tracked into.