Added hints to tell the player to reset if they are walking towards the clouds.
When the player is able to walk on the walls of a puzzleRoom, and end up walking out of the play area towards the clouds, have hints appear to inform the player they should reset/head back.

Also, the color of the hint text inverts when handling input hits, such as running and jumping. This is because the hint text is mostly white, and during those moments where the hint text appears, the area around is mostly white also. This makes the text stick out and is much more visible, since this text is actually important to convey to the player.