Added a colored portals in puzzleRoom 7 (white room)
All the portals in the white room now use a previous room's textures to color coordinate each portal's doorway. Currently, we are only using one texture for all the portals. But, each doorway is applying the proper blending values, so we only need to change the materials for each doorway.