Improved the texture blending in the white puzzle room
The textures that blend between the portals are now more prominent. The textures are now not only used on the outside of the doorframes, but also in the interior, where the portals are placed. A new edge material is used to blend the dark red marble with the white normal marble. This is to make the main portal the player encounters much more unique than the others, making it serve as a landmark.