Updated the white room's texture blending and portal placement
Made a few changes to the white room. Most notably, most of the portals now use edge blending with their textures in relation to their doorways.

Moved the teal portal onto the base floor to make the startingFloor much less empty and give the player more options.

Also, the walking on wall hint text will now not appear in the final room