Simplified the hint system when falling outside and walking across the walls
The hint system has now been simplified on the programming side. It is now much easier to insert new hints in the "sequences" of walking across the puzzleRoom's walls and falling infinitly while outside.

Also, increased the minimum size a camera must be before letting it's portal be rendered as the out of fustrum error was still occuring at times.